Реализовали проверку прав на удаление карточки

+++ controllers/cards.js
@@ -5,8 +5,10 @@ const ObjectID = mongoose.Types.ObjectId;
 const invalidDataMsg = 'Переданы некорректные данные карточки.';
 const cardNotFoundMsg = 'Карточка не найдена.';
 const intServerErrorMsg = 'Внутренняя ошибка сервера.';
+const forbiddenErrorMsg = 'У вас нет прав на этой действие.';
 
 const BAD_REQUEST = 400;
+const FORBIDDEN = 403;
 const NOT_FOUND = 404;
 const INT_SERVER_ERROR = 500;
 
@@ -34,9 +36,19 @@ const deleteCard = (req, res) => {
     res.status(BAD_REQUEST).send({ message: invalidDataMsg });
     return;
   }
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .orFail(() => new Error(cardNotFoundMsg))
-    .then((card) => res.send({ data: card }))
+    .then((card) => {
+      if (card.owner === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((cardToDelete) => res.send({ data: cardToDelete }))
+          .catch(() => {
+            res.status(INT_SERVER_ERROR).send({ message: intServerErrorMsg });
+          });
+      } else {
+        res.status(FORBIDDEN).send({ message: forbiddenErrorMsg });
+      }
+    })
     .catch((err) => {
       if (err.message === cardNotFoundMsg) {
         res.status(NOT_FOUND).send({ message: cardNotFoundMsg });
